requirements.txt
tensorflow==2.20.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.3.0
pillow>=10.0.0
